+++ sites/partners/layouts/index.tsx
@@ -6,13 +6,20 @@ import {
   SiteFooter,
   FooterNav,
   FooterSection,
-  ExygyFooter,
   t,
   AuthContext,
 } from "@bloom-housing/ui-components"
+import { useRouter } from "next/router"
 
 const Layout = (props) => {
   const { profile, signOut } = useContext(AuthContext)
+  const router = useRouter()
+
+  const languages =
+    router?.locales?.map((item) => ({
+      prefix: item === "en" ? "" : item,
+      label: t(`languages.${item}`),
+    })) || []
 
   return (
     <div className="site-wrapper">
@@ -26,6 +33,10 @@ const Layout = (props) => {
           logoSrc="/images/detroit-logo.png"
           notice=""
           title={t("nav.siteTitlePartners")}
+          language={{
+            list: languages,
+            codes: router?.locales,
+          }}
         >
           <LocalizedLink href="/" className="navbar-item">
             {t("nav.listings")}
@@ -62,9 +73,6 @@ const Layout = (props) => {
             <LocalizedLink href="/privacy">{t("pageTitle.privacy")}</LocalizedLink>
             <LocalizedLink href="/disclaimer">{t("pageTitle.disclaimer")}</LocalizedLink>
           </FooterNav>
-          <FooterSection className="bg-black" small>
-            <ExygyFooter />
-          </FooterSection>
         </SiteFooter>
       </div>
     </div>